+++ mne/viz/tests/test_3d.py
@@ -19,9 +19,9 @@ from mne import (make_field_map, pick_channels_evoked, read_evokeds,
 from mne.io import read_raw_ctf, read_raw_bti, read_raw_kit, read_info
 from mne.io.meas_info import write_dig
 from mne.viz import (plot_sparse_source_estimates, plot_source_estimates,
-                     plot_trans, snapshot_brain_montage)
+                     plot_trans, snapshot_brain_montage, plot_head_positions)
 from mne.utils import (requires_mayavi, requires_pysurfer, run_tests_if_main,
-                       _import_mlab, _TempDir)
+                       _import_mlab, _TempDir, requires_nibabel, check_version)
 from mne.datasets import testing
 from mne.source_space import read_source_spaces
 
@@ -49,6 +49,21 @@ config_fname = op.join(base_dir, 'test_config_linux')
 hs_fname = op.join(base_dir, 'test_hs_linux')
 sqd_fname = op.join(io_dir, 'kit', 'tests', 'data', 'test.sqd')
 warnings.simplefilter('always')  # enable b/c these tests throw warnings
+
+
+def test_plot_head_positions():
+    """Test plotting of head positions."""
+    import matplotlib.pyplot as plt
+    pos = np.random.RandomState(0).randn(4, 10)
+    pos[:, 0] = np.arange(len(pos))
+    with warnings.catch_warnings(record=True):  # old MPL will cause a warning
+        plot_head_positions(pos)
+        if check_version('matplotlib', '1.4'):
+            plot_head_positions(pos, mode='field')
+        else:
+            assert_raises(RuntimeError, plot_head_positions, pos, mode='field')
+    assert_raises(ValueError, plot_head_positions, pos, 'foo')
+    plt.close('all')
 
 
 @testing.requires_testing_data
@@ -256,8 +271,32 @@ def test_plot_dipole_locations():
                   subjects_dir, mode='foo')
 
 
+@testing.requires_testing_data
+@requires_nibabel
+def test_plot_dipole_mri_orthoview():
+    """Test mpl dipole plotting."""
+    import matplotlib.pyplot as plt
+    dipoles = read_dipole(dip_fname)
+    trans = read_trans(trans_fname)
+    for coord_frame, idx, show_all in zip(['head', 'mri'],
+                                          ['gof', 'amplitude'], [True, False]):
+        fig = dipoles.plot_locations(trans, 'sample', subjects_dir,
+                                     coord_frame=coord_frame, idx=idx,
+                                     show_all=show_all, mode='orthoview')
+        fig.canvas.scroll_event(0.5, 0.5, 1)  # scroll up
+        fig.canvas.scroll_event(0.5, 0.5, -1)  # scroll down
+        fig.canvas.key_press_event('up')
+        fig.canvas.key_press_event('down')
+        fig.canvas.key_press_event('a')  # some other key
+    ax = plt.subplot(111)
+    assert_raises(ValueError, dipoles.plot_locations, trans, 'sample',
+                  subjects_dir, ax=ax)
+    plt.close('all')
+
+
 @requires_mayavi
 def test_snapshot_brain_montage():
+    """Test snapshot brain montage."""
     info = read_info(evoked_fname)
     fig = plot_trans(info, trans=None, subject='sample',
                      subjects_dir=subjects_dir)
@@ -275,5 +314,6 @@ def test_snapshot_brain_montage():
 
     # Make sure we raise error if the figure has no scene
     assert_raises(TypeError, snapshot_brain_montage, fig, info)
+
 
 run_tests_if_main()